Trim mt_sceneFunc allocation pointer

+++ src/Scene/mt_sceneFunc.c
@@ -4681,7 +4681,7 @@ void FUN_003bceb0(int param_1)
 
 {
 
-  u64 uVar1;
+  u32 uVar1;
 
   int iVar2;
